+++ CotEditor/Sources/CETextSelection.h
@@ -82,6 +82,10 @@ typedef NS_ENUM(NSUInteger, CEUNFType) {
 // AppleScript handler
 - (void)handleShiftRightScriptCommand:(NSScriptCommand *)command;
 - (void)handleShiftLeftScriptCommand:(NSScriptCommand *)command;
+- (void)handleMoveLineUpScriptCommand:(NSScriptCommand *)command;
+- (void)handleMoveLineDownScriptCommand:(NSScriptCommand *)command;
+- (void)handleSortLinesAscendingScriptCommand:(NSScriptCommand *)command;
+- (void)handleReverseLinesScriptCommand:(NSScriptCommand *)command;
 - (void)handleCommentOutScriptCommand:(NSScriptCommand *)command;
 - (void)handleUncommentScriptCommand:(NSScriptCommand *)command;
 - (void)handleChangeCaseScriptCommand:(NSScriptCommand *)command;