+++ CotEditor/Sources/CETextSelection.m
@@ -248,6 +248,42 @@
 
 
 // ------------------------------------------------------
+/// swap selected lines with the line just above
+- (void)handleMoveLineUpScriptCommand:(NSScriptCommand *)command
+// ------------------------------------------------------
+{
+    [[[[self document] editor] focusedTextView] moveLineUp:command];
+}
+
+
+// ------------------------------------------------------
+/// swap selected lines with the line just below
+- (void)handleMoveLineDownScriptCommand:(NSScriptCommand *)command
+// ------------------------------------------------------
+{
+    [[[[self document] editor] focusedTextView] moveLineDown:command];
+}
+
+
+// ------------------------------------------------------
+/// swap selected lines with the line just below
+- (void)handleSortLinesAscendingScriptCommand:(NSScriptCommand *)command
+// ------------------------------------------------------
+{
+    [[[[self document] editor] focusedTextView] sortLinesAscending:command];
+}
+
+
+// ------------------------------------------------------
+/// swap selected lines with the line just below
+- (void)handleReverseLinesScriptCommand:(NSScriptCommand *)command
+// ------------------------------------------------------
+{
+    [[[[self document] editor] focusedTextView] reverseLines:command];
+}
+
+
+// ------------------------------------------------------
 /// comment-out the selection
 - (void)handleCommentOutScriptCommand:(NSScriptCommand *)command
 // ------------------------------------------------------