+++ CotEditor/Sources/CETextView.h
@@ -114,3 +114,14 @@
 - (IBAction)editColorCode:(nullable id)sender;
 
 @end
+
+
+@interface CETextView (LineProcessing)
+
+- (IBAction)moveLineUp:(nullable id)sender;
+- (IBAction)moveLineDown:(nullable id)sender;
+- (IBAction)sortLinesAscending:(nullable id)sender;
+- (IBAction)reverseLines:(nullable id)sender;
+- (IBAction)deleteLine:(nullable id)sender;
+
+@end